Python/requirements.txt
opencv-python>=4.8.0
ultralytics>=8.0.0
paho-mqtt>=2.0.0
mysql-connector-python>=8.0.0
numpy>=1.21.0
threading-timer>=1.0.0 